% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdx.R
\name{all_items.gdx}
\alias{all_items.gdx}
\title{Return the list of all items}
\usage{
\method{all_items}{gdx}(x)
}
\arguments{
\item{x}{the gdx object}
}
\description{
Return the list of all items
}
\examples{
\dontrun{
    mygdx <- gdx('results.gdx')
    all_items(mygdx)
 }
}

