% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{all_items}
\alias{all_items}
\title{Return the list of all items}
\usage{
all_items(x, ...)
}
\arguments{
\item{x}{the gdx object}

\item{...}{ignored}
}
\value{
A named \code{list} with four character vectors of symbol names:
  \code{variables}, \code{parameters}, \code{sets} and \code{equations}.
}
\description{
Return the list of all items
}
\author{
Laurent Drouet
}
