% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdxtools.R
\name{batch_extract}
\alias{batch_extract}
\title{Extract a list of items from many GDX}
\usage{
batch_extract(items, files = NULL, gdxs = NULL, ...)
}
\arguments{
\item{items}{vector or list of items to extract}

\item{files}{list of files; if \code{NULL}, \code{gdxs} must be supplied}

\item{gdxs}{list of \code{gdx} objects; if \code{NULL}, \code{files} must be supplied}

\item{...}{passed to \code{extract.gdx}}
}
\value{
A named \code{list} with one \code{data.frame} per requested item,
  each the row-bind of that item extracted from every file with a
  \code{gdx} column identifying the source file.
}
\description{
Extract a list of items from many GDX
}
\examples{
 f1 <- tempfile(fileext = ".gdx")
 f2 <- tempfile(fileext = ".gdx")
 write.gdx(f1, list(myparam = data.frame(i = c("a", "b"), value = 1:2)))
 write.gdx(f2, list(myparam = data.frame(i = c("a", "b"), value = 3:4)))
 allparam <- batch_extract("myparam", files = c(f1, f2))
}
\author{
Laurent Drouet
}
