% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dtsample}
\alias{dtsample}
\title{dtsample}
\usage{
dtsample(table, size = 10)
}
\arguments{
\item{table}{a data.table.}

\item{size}{sample size.}
}
\value{
a data.table of length `size`.
}
\description{
Sample a data.table.
}
\examples{
dtsample(table, 10)
}
