% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{extract}
\alias{extract}
\title{Extract data from a gdx}
\usage{
extract(x, ...)
}
\arguments{
\item{x}{the gdx object}

\item{...}{arguments passed to \code{extract.gdx}}
}
\value{
A \code{data.frame}; see \code{\link{extract.gdx}} for the column
  layout and attributes.
}
\description{
Extract data from a gdx
}
