% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdxrrw.R
\name{gams}
\alias{gams}
\title{Run the GAMS executable}
\usage{
gams(gmsAndArgs)
}
\arguments{
\item{gmsAndArgs}{command-line string passed to \code{gams}}
}
\value{
exit status (integer)
}
\description{
Thin wrapper around \code{system2("gams", ...)}. Returns the integer exit
status from GAMS.
}
