% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{gdx}
\alias{gdx}
\title{GDX file}
\usage{
gdx(filename, ...)
}
\arguments{
\item{filename}{filename of the gdx file}

\item{...}{extra fields stored on the resulting object}
}
\description{
Constructs a \code{gdx} object backed by a \code{gamstransfer::Container}.
The container loads the GDX once and is reused on subsequent
\code{extract}/\code{[} calls.
}
\examples{
 \dontrun{
   mygdx <- gdx('results.gdx')
 }
}
\author{
Laurent Drouet
}
