% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx_gdxrrw.R
\name{gdx_gdxrrw}
\alias{gdx_gdxrrw}
\title{GDX file}
\usage{
gdx_gdxrrw(filename, ...)
}
\arguments{
\item{filename}{filename of the gdx file}
}
\description{
GDX file
}
\examples{

  \dontrun{
    mygdx <- gdx('results.gdx')
  }
}
