% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{raw2gdx}
\alias{raw2gdx}
\title{raw2gdx}
\usage{
raw2gdx(gdx, var)
}
\arguments{
\item{gdx}{the gdx filename.}

\item{var}{list of properties of a gdx symbol as provided by gdxrrw::rgdx.}
}
\description{
Save to a GAMS gdx file. Works on a named list providing
domains and data as given by gdxrrw::rgdx.
}
\examples{
# round-trip:
raw2gdx("gdxname.gdx", gdxrrw::rgdx("gdxname.gdx", list(name="symbolname"))
}
