% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_variable}
\alias{read_variable}
\title{readVariable}
\usage{
read_variable(fname, varname, field = "l")
}
\arguments{
\item{fname}{the gdx filename.}

\item{varname}{name of the variable to load.}

\item{field}{select a field or use "all".}
}
\value{
a data.table
}
\description{
Read a variable from a gdx file.
}
\examples{
readVariable("fulldata.gdx", "v35_shEsPeT", "l")
}
