% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdxrrw.R
\name{rgdx}
\alias{rgdx}
\title{include some parts of gdxrrw from https://support.gams.com/gdxrrw:interfacing_gams_and_r}
\usage{
rgdx(gdxName, requestList = NULL, squeeze = TRUE, useDomInfo = TRUE,
  followAlias = TRUE)
}
\description{
include some parts of gdxrrw from https://support.gams.com/gdxrrw:interfacing_gams_and_r
}
\author{
Steve Dirkse
}
