% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{[.gdx}
\alias{[.gdx}
\title{Extract data from a gdx}
\usage{
\method{[}{gdx}(x, ...)
}
\description{
Extract data from a gdx
}
\examples{

  \dontrun{
    mygdx <- gdx('results.gdx')
    travel_cost <- mygdx["travel_cost"]
  }
}
\author{
Laurent Drouet
}
