% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{years_as_numbers}
\alias{years_as_numbers}
\title{years_as_numbers}
\usage{
years_as_numbers(table)
}
\arguments{
\item{table}{a data.table.}
}
\value{
a data.table
}
\description{
Specific to data provided by the REMIND IAM.
Coerces values in "time-like" columns in a data.table to numbers.
The provided data.table is modified *in-place*.
}
\examples{
years_as_numbers(table)
}
