% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdxtools.R
\name{batch_extract}
\alias{batch_extract}
\title{Extract a list of items from many GDX}
\usage{
batch_extract(items, files = NULL, gdxs = NULL)
}
\arguments{
\item{items}{vector or list of items to extract}

\item{files}{list of files, if not defined gdxs should be defined}

\item{gdxs}{list of gdxs, if not defined files should be defined}
}
\description{
Extract a list of items from many GDX
}
\examples{

  \dontrun{
    myfiles = c("test1.gdx","test2.gdx")
    allparam <- batch_extract("myparam",files=myfiles)
  }
}
\author{
Laurent Drouet
}
