% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{extract.gdx}
\alias{extract.gdx}
\title{Extract parameter or variable data from the gdx file}
\usage{
\method{extract}{gdx}(x, item, field = "l", addgdx = FALSE, ...)
}
\arguments{
\item{x}{the gdx object}

\item{item}{the name of the item to extract}

\item{field}{the field of the variable to be extracted. Can be 'l', 'm',
'lo', 'up' (level, marginal, lower, upper). Defaults to level.}

\item{addgdx}{if \code{TRUE}, append a \code{gdx} column with the filename.}

\item{...}{ignored; for backward compatibility.}
}
\description{
Extract parameter or variable data from the gdx file
}
\examples{
 \dontrun{
   mygdx <- gdx('results.gdx')
   travel_cost <- mygdx["travel_cost"]
   travel_cost <- extract(mygdx, "travel_cost")
 }
}
\author{
Laurent Drouet
}
