% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{extract.gdx}
\alias{extract.gdx}
\title{Extract parameter or variable data from the gdx file}
\usage{
\method{extract}{gdx}(x, item, field = "l", addgdx = F, ...)
}
\arguments{
\item{x}{the gdx object}

\item{item}{the name of the item to extract}

\item{field}{the field of the variable to be extracted. Can be 'l', 'm', 'lo'
'up', respectively for level, marginal, lower bound, upper bound.
Defaults to level.}

\item{addgdx}{if \code{TRUE}, the data.frame will get an extra column
with the filename of the gdx.}
}
\description{
Extract parameter or variable data from the gdx file
}
\examples{
 \dontrun{
    mygdx <- gdx('results.gdx')
    travel_cost <- mygdx["travel_cost"]
    travel_cost <- extract(mygdx,"travel_cost")
 }

}
\author{
Laurent Drouet
}
