% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx.R
\name{gdx}
\alias{gdx}
\title{GDX file}
\usage{
gdx(filename, ...)
}
\arguments{
\item{filename}{filename of the gdx file}
}
\description{
GDX file
}
\examples{

  \dontrun{
    mygdx <- gdx('results.gdx')
  }
}
\author{
Laurent Drouet
}
