% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdxtools.R
\name{write.gdx}
\alias{write.gdx}
\title{Write a list of parameters in a gdx}
\usage{
write.gdx(file, params = list(), removeLST = T, usetempdir = T)
}
\arguments{
\item{file}{the filename of the gdx to save}

\item{params}{named list of parameters}

\item{removeLST}{remove temporary lst file}

\item{usetempdir}{uses system temp dir for the temporary files, otherwise use local file "tmp.gms"}
}
\description{
Write a list of parameters in a gdx
}
\examples{
 \dontrun{
    param1 = data.frame(x=c('1','2'),value=1:10)
    param2 = data.frame(a=c('london','paris','tahiti'),value=c(50,0.2,1e-2))
    write.gdx("test.gdx",list(param1=param1,param2=param2))
 }

}

