% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdxtools.R
\name{write.gdx}
\alias{write.gdx}
\title{Write a list of parameters in a gdx (old version which uses a temporary gams file)}
\usage{
write.gdx(file, params = list(), vars_l = list(), vars_lo = list(),
  vars_up = list(), sets = list(), removeLST = T, usetempdir = T,
  digits = 16, compress = F)
}
\arguments{
\item{file}{the filename of the gdx to save}

\item{params}{named list of parameters}

\item{vars_l}{named list of variable levels}

\item{vars_lo}{named list of variable lower bounds}

\item{vars_up}{named list of variable upper bounds}

\item{sets}{named list of sets}

\item{removeLST}{remove temporary lst file}

\item{usetempdir}{uses system temp dir for the temporary files, otherwise use local file "tmp.gms"}

\item{compress}{compress GDX}

\item{digit}{number of digits to use}
}
\description{
Write a list of parameters in a gdx (old version which uses a temporary gams file)
}
\examples{
 \dontrun{
    param1 = data.frame(x=c('1','2'),value=1:10)
    param2 = data.frame(a=c('london','paris','tahiti'),value=c(50,0.2,1e-2))
    write.gdx("test.gdx",list(param1=param1,param2=param2))
 }

}
\author{
Laurent Drouet
}
