% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdxtools.R
\name{write2.gdx}
\alias{write2.gdx}
\title{Write a list of parameters or sets in a gdx, using gdxrrw (faster than write.gdx, but less options)}
\usage{
write2.gdx(file, params = list(), sets = list())
}
\arguments{
\item{file}{the filename of the gdx to save}

\item{params}{named list of parameters}

\item{sets}{named list of sets}
}
\description{
Write a list of parameters or sets in a gdx, using gdxrrw (faster than write.gdx, but less options)
}
\examples{
 \dontrun{
    param1 = data.frame(x=c('1','2'),value=1:10)
    param2 = data.frame(a=c('london','paris','tahiti'),value=c(50,0.2,1e-2))
    write.gdx("test.gdx",list(param1=param1,param2=param2))
 }

}
\author{
Laurent Drouet
}
